refactor(template-compiler): cleanup imports
